% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aceecostats-package.r
\name{aes_maxdate aes_decades}
\alias{aes_maxdate aes_decades}
\title{The key dates used for ACE ecostats assessment in the latest summary.}
\description{
\code{aes_maxdata} is the final date used for inclusion 
\code{aes_decades} is the break points to defined periods
}

