% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.r
\name{aes_season}
\alias{aes_season}
\title{Season classifying}
\usage{
aes_season(x)
}
\arguments{
\item{x}{date}
}
\value{
text Spring, Summer, Autumn, or Winter
}
\description{
Season classifying
}
\examples{
aes_season(Sys.Date())
}

