% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aceecostats-package.r
\name{aes_zone}
\alias{aes_zone}
\alias{aes_zone_ll}
\title{Assessment zones, simplified and made clear for Sector x Zone.}
\description{
Zones are four sectors (ocean basins Atlantic, Indian, East Pacific, West Pacific) and three zones (Mid-Latitude, High-Latitude
and Continent (Antarctic-shelf)).
}
\examples{
library(sp)
op <- par(mfrow = c(2, 1), mar = c(0,0, 0,0))
plot(aes_zone_ll, asp = 1.7); text(coordinates(aes_zone_ll), lab = 1:12)
plot(aes_zone); text(coordinates(aes_zone), lab = 1:12)
par(op)
}
