% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_data_prep.r
\name{build_bulk_file}
\alias{build_bulk_file}
\title{File-backed raster file.}
\usage{
build_bulk_file(files, filename, read_i_raster = NULL, layer_prefix = "lyr")
}
\arguments{
\item{files}{data frame with `fullname` and `date`}

\item{filename}{database of files, see Details}

\item{read_i_raster}{function to read the i-th file}

\item{layer_prefix}{optional prefix to give each layer, pre-pended to the date}
}
\value{
RasterBrick
}
\description{
Build a file-backed raster from a collection of file names and dates.
}
\details{
The files collection must have columns \code{fullname} with the path to the file, and \code{date} with a valid date or date-time.
The function \code{read_i_raster} takes an integer and the files collection and returns the raster read from the i-th row.
The date is set on the object after this function returns, to ensure it matches the database of files, and also so that the raster function can be used as a fallback.
}

