% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_cells_polygon.R
\name{classify_cells_by_polygon}
\alias{classify_cells_by_polygon}
\title{Classify raster cells by polygons}
\usage{
classify_cells_by_polygon(cells, grid, polygons)
}
\arguments{
\item{cells}{raster cell index}

\item{grid}{raster for the cells}

\item{polygons}{SpatialPolygonsDataFrame}
}
\description{
from polygons, copy the first column of data for every cell
the cells and grid MUST match, but the polygons can be in any projection
Make sure the polygons have the first column as the classifier you want
}
