% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_data_prep.r
\name{decade_maker}
\alias{decade_maker}
\title{Decade classification}
\usage{
decade_maker(x)
}
\arguments{
\item{x}{date-time vector}
}
\value{
classified factor vector
}
\description{
This changes over time, but currently it's clean on 10 year boundaries (1977-2017).
}
