% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker-plot.r
\name{do_hist}
\alias{do_hist}
\title{Worker function for histogram/density}
\usage{
do_hist(v, w = NULL)
}
\arguments{
\item{v}{measured values (e.g. ice season duration)}

\item{w}{weight values (e.g. area of this cell)}
}
\description{
Worker function for histogram/density
}

