% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-ice-season.R
\name{readIceBrick}
\alias{readIceBrick}
\title{function to read a series of time slices, and interpolate a full sequence optionally}
\usage{
readIceBrick(dates, hemisphere = "south", interpolate = TRUE)
}
\description{
function to read a series of time slices, and interpolate a full sequence optionally
}
