% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-ice-season.R
\name{readIceCache}
\alias{readIceCache}
\title{Read ice cache
 
 function to read a series of time slices, and interpolate a full sequence of days
 
 this version of readIceBrick replaces it to get it from ice.grd}
\usage{
readIceCache(dates, filename)
}
\arguments{
\item{dates}{dates to return}

\item{filename}{bulk file source}
}
\description{
Read ice cache
 
 function to read a series of time slices, and interpolate a full sequence of days
 
 this version of readIceBrick replaces it to get it from ice.grd
}
