% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker-plot.r
\name{sector_name}
\alias{sector_name}
\title{Sector name/label mapping}
\usage{
sector_name(secname)
}
\arguments{
\item{secname}{short name}
}
\value{
the long names
}
\description{
Sector name/label mapping
}

