% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-ice-season.R
\name{snaptodoy}
\alias{snaptodoy}
\title{function to ensure we have clean boundaries}
\usage{
snaptodoy(date, mon = 2, mday = 15, start = TRUE)
}
\description{
function to ensure we have clean boundaries
}
