% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{tabit}
\alias{tabit}
\title{Convert a raster to a data frame.}
\usage{
tabit(x, na.rm = TRUE)
}
\arguments{
\item{x}{RasterLayer}

\item{na.rm}{remove missing values}
}
\description{
Create a data frame from a single layer [raster::RasterLayer()]. This is
}

