% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aceecostats-package.r
\name{aes_files}
\alias{aes_chlfiles}
\alias{aes_files}
\alias{aes_icefiles}
\alias{aes_sstfiles}
\title{The available remote sensing files at the time of bulk extraction.}
\description{
The available remote sensing files at the time of bulk extraction.
}

